django_passwords/requirements.txt
# Django core
Django==5.1.5
django-environ==0.12.0
django-crispy-forms==2.3
crispy-bootstrap5==2024.10
django-extensions==3.2.3
django-formtools==2.5.1
django-auto-logout==0.5.1
django-axes==7.0.2
django-otp==1.5.4
django-two-factor-auth==1.17.0
django-recaptcha==4.0.0
django-phonenumber-field==8.0.0
phonenumbers==8.13.53
python-dotenv==1.0.1
psycopg2-binary==2.9.10
qrcode==7.4.2

# Machine Learning / Deep Learning
numpy==2.0.2
pandas==2.2.1
scikit-learn==1.5.0
matplotlib==3.8.3
seaborn==0.13.2
scipy==1.13.1
tensorflow==2.18.0
keras==3.8.0
xgboost==2.0.3
lightgbm==4.3.0
optuna==3.5.0

# Herramientas adicionales
joblib==1.4.2
statsmodels==0.14.1
pydantic==2.10.5

# Integración con JS / Frontend moderno (API, WebSockets, etc.)
djangorestframework==3.15.1
channels==4.1.0
